feat: 2주차 미션 파일 이름 수정, 뷰바인등으로 수정
