+++ potato/two_week/app/src/main/java/com/cookandroid/two_week/MainActivity.kt
@@ -3,34 +3,29 @@ package com.cookandroid.two_week
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.cookandroid.two_week.databinding.AcitivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
+    // 뷰 바인딩 객체 선언
+    private lateinit var binding: AcitivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.acitivity_main)
+
+        // 뷰 바인딩 객체 초기화
+        binding = AcitivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
         supportActionBar?.hide() // 액티비티 제목 숨기기
-
-        // 예시로 Fragment1에서 Fragment2로 이동할 때 슬라이드 효과 추가
-        /*val fragmentTransaction = supportFragmentManager.beginTransaction()
-
-        // 슬라이드 애니메이션
-        fragmentTransaction.setCustomAnimations(
-            android.R.anim.slide_in_left,   // 화면에 등장할 때 애니메이션
-            android.R.anim.slide_out_right  // 화면에서 사라질 때 애니메이션
-        )*/
-
-
-
-        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottom_navigation)
 
         // 앱 실행 시 기본으로 표시할 Fragment 설정
         if (savedInstanceState == null) {
             replaceFragment(HomeFragment())
         }
 
-        bottomNavigationView.setOnItemSelectedListener { item ->
+        // BottomNavigationView 클릭 리스너 설정
+        binding.bottomNavigation.setOnItemSelectedListener { item ->
             when (item.itemId) {
                 R.id.nav_home -> replaceFragment(HomeFragment())
                 R.id.nav_shopping -> replaceFragment(ShoppingFragment())
@@ -48,7 +43,5 @@ class MainActivity : AppCompatActivity() {
             .setCustomAnimations(android.R.anim.fade_in, android.R.anim.fade_out) // 애니메이션 추가
             .replace(R.id.fragment_container, fragment) // 새로운 Fragment로 교체
             .commit() // 변경 사항 적용
-        //페이드 아웃(fade out) 효과를 적용함.
-
     }
 }
