--- potato/two_week/app/src/main/java/com/cookandroid/two_week/Shopping.kt
@@ -1,16 +1,0 @@
-package com.cookandroid.two_week
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class ShoppingFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return inflater.inflate(R.layout.fragment_shopping, container, false)
-    }
-}
