+++ potato/two_week/app/src/main/java/com/cookandroid/two_week/ShoppingFragment.kt
@@ -1,0 +1,27 @@
+package com.cookandroid.two_week
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.cookandroid.two_week.databinding.FragmentShoppingBinding
+
+class ShoppingFragment : Fragment() {
+
+    private var _binding: FragmentShoppingBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentShoppingBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
